Move code that reads landmarks splitting borders into LM handler
