+++ core/src/main/java/com/graphhopper/routing/lm/LMPreparationHandler.java
@@ -17,25 +17,33 @@
  */
 package com.graphhopper.routing.lm;
 
+import com.bedatadriven.jackson.datatype.jts.JtsModule;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.graphhopper.GraphHopperConfig;
 import com.graphhopper.config.LMProfile;
 import com.graphhopper.routing.ch.CHPreparationHandler;
+import com.graphhopper.routing.util.spatialrules.AbstractSpatialRule;
+import com.graphhopper.routing.util.spatialrules.SpatialRuleLookup;
+import com.graphhopper.routing.util.spatialrules.SpatialRuleLookupBuilder;
 import com.graphhopper.storage.GraphHopperStorage;
 import com.graphhopper.storage.StorableProperties;
 import com.graphhopper.storage.index.LocationIndex;
+import com.graphhopper.util.JsonFeatureCollection;
 import com.graphhopper.util.Parameters;
 import com.graphhopper.util.Parameters.Landmark;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.FileInputStream;
 import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.Reader;
 import java.util.*;
 import java.util.concurrent.ExecutorCompletionService;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-import static com.graphhopper.util.Helper.createFormatter;
-import static com.graphhopper.util.Helper.getMemInfo;
+import static com.graphhopper.util.Helper.*;
 
 /**
  * This class deals with the A*, landmark and triangulation (ALT) preparations.
@@ -57,6 +65,7 @@ public class LMPreparationHandler {
     private int preparationThreads;
     private ExecutorService threadPool;
     private boolean logDetails = false;
+    private SpatialRuleLookup ruleLookup;
 
     public LMPreparationHandler() {
         setPreparationThreads(1);
@@ -78,6 +87,22 @@ public class LMPreparationHandler {
         for (String loc : ghConfig.getString(Landmark.PREPARE + "suggestions_location", "").split(",")) {
             if (!loc.trim().isEmpty())
                 lmSuggestionsLocations.add(loc.trim());
+        }
+
+        if (!isEnabled())
+            return;
+
+        String splitAreaLocation = ghConfig.getString(Landmark.PREPARE + "split_area_location", "");
+        JsonFeatureCollection landmarkSplittingFeatureCollection = loadLandmarkSplittingFeatureCollection(splitAreaLocation);
+        if (landmarkSplittingFeatureCollection != null && !landmarkSplittingFeatureCollection.getFeatures().isEmpty()) {
+            ruleLookup = SpatialRuleLookupBuilder.buildIndex(
+                    Collections.singletonList(landmarkSplittingFeatureCollection), "area",
+                    (id, polygons) -> new AbstractSpatialRule(polygons) {
+                        @Override
+                        public String getId() {
+                            return id;
+                        }
+                    });
         }
     }
 
@@ -186,6 +211,12 @@ public class LMPreparationHandler {
      * @see CHPreparationHandler#prepare(StorableProperties, boolean) for a very similar method
      */
     public boolean loadOrDoWork(final StorableProperties properties, final boolean closeEarly) {
+        for (PrepareLandmarks prep : preparations) {
+            // the ruleLookup splits certain areas from each other but avoids making this a permanent change so that other algorithms still can route through these regions.
+            if (ruleLookup != null && !ruleLookup.getRules().isEmpty()) {
+                prep.setSpatialRuleLookup(ruleLookup);
+            }
+        }
         ExecutorCompletionService<String> completionService = new ExecutorCompletionService<>(threadPool);
         int counter = 0;
         final AtomicBoolean prepared = new AtomicBoolean(false);
@@ -258,6 +289,21 @@ public class LMPreparationHandler {
             if (minNodes > 1)
                 tmpPrepareLM.setMinimumNodes(minNodes);
             addPreparation(tmpPrepareLM);
+        }
+    }
+
+    private JsonFeatureCollection loadLandmarkSplittingFeatureCollection(String splitAreaLocation) {
+        ObjectMapper objectMapper = new ObjectMapper();
+        objectMapper.registerModule(new JtsModule());
+        try (Reader reader = splitAreaLocation.isEmpty() ?
+                new InputStreamReader(LandmarkStorage.class.getResource("map.geo.json").openStream(), UTF_CS) :
+                new InputStreamReader(new FileInputStream(splitAreaLocation), UTF_CS)) {
+            JsonFeatureCollection result = objectMapper.readValue(reader, JsonFeatureCollection.class);
+            LOGGER.info("Loaded landmark splitting collection from " + splitAreaLocation);
+            return result;
+        } catch (IOException e) {
+            LOGGER.error("Problem while reading border map GeoJSON. Skipping this.", e);
+            return null;
         }
     }
 }
